break out snakemake tests
begin process of separating out all tests of snakemake pipelines from
unit and integration into a separate space & Travis job.
